+++ apps/partners-reference/pages/add.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import Head from "next/head"
 import { PageHeader, t, MetaTags } from "@bloom-housing/ui-components"
 import Layout from "../layouts/application"
-import ApplicationForm from "../src/applications/ApplicationForm/ApplicationForm"
+import PaperApplicationForm from "../src/applications/PaperApplicationForm/PaperApplicationForm"
 
 const NewApplication = () => {
   const metaDescription = ""
@@ -16,7 +16,7 @@ const NewApplication = () => {
       <MetaTags title={t("nav.siteTitle")} image={metaImage} description={metaDescription} />
       <PageHeader>{t("applications.newApplication")}</PageHeader>
 
-      <ApplicationForm isEditable={false} />
+      <PaperApplicationForm />
     </Layout>
   )
 }
