+++ apps/partners-reference/src/applications/PaperApplicationForm/sections/FormTerms.tsx
